Mypy fixes

+++ Parser/RobotDataParser.py
@@ -4,7 +4,7 @@ from typing import Any
 class RobotDataParser:
     @staticmethod
     def convert_sensor_data_to_dict(
-        data: list[tuple[float, float]]
+        data: list[tuple[int, float]]
     ) -> list[dict[str, Any]]:
         return [{"angle": reading[1], "value": reading[0]} for reading in data]
